Add state for count of items in cart and pass to children
Create state variable for count and event handlers for adding and
subtracting. Pass the state to the Navbar component where it will
eventually be displayed with the cart icon. Pass the state and the
event handlers as context to the Outlet that shows the remaining
pages and be able to increment/decrement the parent's state variable
from multiple pages.